Update merge TokenModule into UtilModule
