--- libs/token/src/index.ts
@@ -1,3 +1,0 @@
-export * from './token.module';
-export * from './token.service';
-export * from './token.class';
