--- libs/token/tslint.json
@@ -1,6 +1,0 @@
-{
-    "extends": "../../tslint.json",
-    "rules": {
-        "variable-name": false
-    }
-}
