+++ libs/util/src/class/token.class.ts
@@ -1,6 +1,6 @@
 export class Token {
-  public readonly accessToken: string;
-  public readonly refreshToken: string;
+  public readonly accessToken?: string;
+  public readonly refreshToken?: string;
 
   constructor(token) {
     Object.assign(this, token);
