+++ nest-cli.json
@@ -29,15 +29,6 @@
         "tsConfigPath": "libs/user/tsconfig.lib.json"
       }
     },
-    "token": {
-      "type": "library",
-      "root": "libs/token",
-      "entryFile": "index",
-      "sourceRoot": "libs/token/src",
-      "compilerOptions": {
-        "tsConfigPath": "libs/token/tsconfig.lib.json"
-      }
-    },
     "restaurant": {
       "type": "library",
       "root": "libs/restaurant",
