+++ package.json
@@ -82,8 +82,6 @@
       "@app/mongo": "<rootDir>/libs/mongo/src",
       "@app/user/(.*)": "<rootDir>/libs/user/src/$1",
       "@app/user": "<rootDir>/libs/user/src",
-      "@app/token/(.*)": "<rootDir>/libs/token/src/$1",
-      "@app/token": "<rootDir>/libs/token/src",
       "@app/restaurant/(.*)": "<rootDir>/libs/rastaurant/src/$1",
       "@app/restaurant": "<rootDir>/libs/rastaurant/src",
       "@app/util/(.*)": "<rootDir>/libs/util/src/$1",
